Legg filtrerUtUtbetalingsperioderMedSammeDatoSomAvslagsperioder og filtrerUtPerioderBasertPåEndringstidspunkt i utilfunksjoner
